+++ pkg/odo/cli/service/operator_backend.go
@@ -11,6 +11,8 @@ import (
 	"os"
 	"strings"
 
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+
 	"github.com/ghodss/yaml"
 	"github.com/openshift/odo/pkg/log"
 	"github.com/openshift/odo/pkg/service"
@@ -53,7 +55,7 @@ func (b *OperatorBackend) CompleteServiceCreate(o *CreateOptions, cmd *cobra.Com
 }
 
 func (b *OperatorBackend) ValidateServiceCreate(o *CreateOptions) (err error) {
-	d := svc.NewDynamicCRD()
+	u := unstructured.Unstructured{}
 	// if the user wants to create service from a file, we check for
 	// existence of file and validate if the requested operator and CR
 	// exist on the cluster
@@ -68,50 +70,26 @@ func (b *OperatorBackend) ValidateServiceCreate(o *CreateOptions) (err error) {
 			return err
 		}
 
-		err = yaml.Unmarshal(fileContents, &d.OriginalCRD)
+		err = yaml.Unmarshal(fileContents, &u.Object)
 		if err != nil {
 			return err
 		}
 
-		// Check if the operator and the CR exist on cluster
-		var csv olm.ClusterServiceVersion
-		b.CustomResource, csv, err = svc.GetCSV(o.KClient, d.OriginalCRD)
-		if err != nil {
-			return err
-		}
+		gvk := svc.GetGVKFromUnstructured(u)
+		b.group, b.version, b.kind = gvk.Group, gvk.Version, gvk.Kind
 
-		// all is well, let's populate the fields required for creating operator backed service
-		b.group, b.version, b.resource, err = svc.GetGVRFromOperator(csv, b.CustomResource)
-		if err != nil {
-			return err
-		}
-
-		err = d.ValidateMetadataInCRD()
-		if err != nil {
-			return err
+		if u.GetName() == "" {
+			return fmt.Errorf("couldn't find \"metadata.name\" in the yaml; provide a name for the service")
 		}
 
 		if o.ServiceName != "" && !o.DryRun {
-			// First check if service with provided name already exists
-			svcFullName := strings.Join([]string{b.CustomResource, o.ServiceName}, "/")
-			exists, err := svc.OperatorSvcExists(o.KClient, svcFullName)
-			if err != nil {
-				return err
-			}
-			if exists {
-				return fmt.Errorf("service %q already exists; please provide a different name or delete the existing service first", svcFullName)
-			}
-
-			d.SetServiceName(o.ServiceName)
+			u.SetName(o.ServiceName)
 		} else {
-			o.ServiceName, err = d.GetServiceNameFromCRD()
-			if err != nil {
-				return err
-			}
+			o.ServiceName = u.GetName()
 		}
 
 		// CRD is valid. We can use it further to create a service from it.
-		b.CustomResourceDefinition = d.OriginalCRD
+		b.CustomResourceDefinition = u.Object
 
 		return nil
 	} else if b.CustomResource != "" {
@@ -138,14 +116,14 @@ func (b *OperatorBackend) ValidateServiceCreate(o *CreateOptions) (err error) {
 				return err
 			}
 
-			d.OriginalCRD = builtCRD
+			u.Object = builtCRD
 		} else {
 			almExample, err := svc.GetAlmExample(csv, b.CustomResource, o.ServiceType)
 			if err != nil {
 				return err
 			}
 
-			d.OriginalCRD = almExample
+			u.Object = almExample
 		}
 
 		if o.ServiceName != "" && !o.DryRun {
@@ -158,23 +136,18 @@ func (b *OperatorBackend) ValidateServiceCreate(o *CreateOptions) (err error) {
 			if exists {
 				return fmt.Errorf("service %q already exists; please provide a different name or delete the existing service first", svcFullName)
 			}
+			u.SetName(o.ServiceName)
 		}
 
-		d.SetServiceName(o.ServiceName)
-
-		err = d.ValidateMetadataInCRD()
-		if err != nil {
-			return err
+		if u.GetName() == "" {
+			return fmt.Errorf("couldn't find metadata.name in the yaml; provide a name for the service")
 		}
 
 		// CRD is valid. We can use it further to create a service from it.
-		b.CustomResourceDefinition = d.OriginalCRD
+		b.CustomResourceDefinition = u.Object
 
 		if o.ServiceName == "" {
-			o.ServiceName, err = d.GetServiceNameFromCRD()
-			if err != nil {
-				return err
-			}
+			o.ServiceName = u.GetName()
 		}
 
 		return nil